+++ packages/obra-icons-svelte/src/package/icons/Avocado.svelte
@@ -16,7 +16,7 @@
 	xmlns="http://www.w3.org/2000/svg"
 	class="obra-icon"
 >
-	<g class="oi-avocado" clip-path="url(#clip0_663_99055)">
+	<g class="oi-avocado" clip-path="url(#clip0_664_159924)">
 		<circle
 			class="oi-ellipse"
 			cx="12"
@@ -33,7 +33,7 @@
 		/>
 	</g>
 	<defs>
-		<clipPath class="clip0_663_99055">
+		<clipPath class="clip0_664_159924">
 			<rect width={size} height={size} fill="white" />
 		</clipPath>
 	</defs>