+++ packages/obra-icons-svelte/src/package/icons/Dress.svelte
@@ -16,7 +16,7 @@
 	xmlns="http://www.w3.org/2000/svg"
 	class="obra-icon"
 >
-	<g class="oi-dress" clip-path="url(#clip0_663_100583)">
+	<g class="oi-dress" clip-path="url(#clip0_664_161432)">
 		<path
 			class="oi-line"
 			d="M15 3.30188V0.75"
@@ -51,7 +51,7 @@
 		/>
 	</g>
 	<defs>
-		<clipPath class="clip0_663_100583">
+		<clipPath class="clip0_664_161432">
 			<rect width={size} height={size} fill="white" />
 		</clipPath>
 	</defs>