+++ packages/obra-icons-svelte/src/package/icons/PreviousFill.svelte
@@ -15,7 +15,7 @@
 	xmlns="http://www.w3.org/2000/svg"
 	class="obra-icon"
 >
-	<g class="oi-previous-fill" clip-path="url(#clip0_663_99712)">
+	<g class="oi-previous-fill" clip-path="url(#clip0_664_160566)">
 		<path
 			class="oi-fill"
 			fill-rule="evenodd"
@@ -35,7 +35,7 @@
 		</g>
 	</g>
 	<defs>
-		<clipPath class="clip0_663_99712">
+		<clipPath class="clip0_664_160566">
 			<rect width={size} height={size} fill="white" />
 		</clipPath>
 	</defs>