+++ packages/obra-icons-svelte/src/package/icons/SneakerMove.svelte
@@ -27,7 +27,7 @@
 		/>
 		<path
 			class="oi-vector"
-			d="M9.12284 10.5L5.02502 6.40125"
+			d="M9.12281 10.5L5.025 6.40125"
 			stroke={color}
 			stroke-width={strokeWidth}
 			stroke-linecap="round"