+++ packages/obra-icons-svelte/src/package/icons/SolarPanel.svelte
@@ -27,7 +27,7 @@
 		/>
 		<path
 			class="oi-line"
-			d="M6.16595 3.91595L7.2272 4.9772"
+			d="M6.16589 3.91595L7.22714 4.9772"
 			stroke={color}
 			stroke-width={strokeWidth}
 			stroke-linecap="round"
@@ -75,7 +75,7 @@
 		/>
 		<path
 			class="oi-line"
-			d="M4.36969 16.5H19.6303"
+			d="M4.36963 16.5H19.6303"
 			stroke={color}
 			stroke-width={strokeWidth}
 			stroke-linecap="round"