+++ packages/obra-icons-svelte/src/package/icons/Volume2.svelte
@@ -19,7 +19,7 @@
 	<g class="oi-volume-2">
 		<path
 			class="oi-vector"
-			d="M10.9999 5L5.99989 9H1.99989V15H5.99989L10.9999 19V5Z"
+			d="M10.9999 5L5.9999 9H1.9999V15H5.9999L10.9999 19V5Z"
 			stroke={color}
 			stroke-width={strokeWidth}
 			stroke-linecap="round"